requirements.txt
pydub==0.25.1
python-swiftclient==3.10.1
ffmpeg-python==0.2.0
uvicorn==0.17.6
python-dotenv==0.19.2
gunicorn==20.1.0
fastapi==0.75.0